Created inform action

+++ spec/factories/events.rb
@@ -4,7 +4,7 @@ FactoryBot.define do
   factory :event do
     name { Faker::Music.band + ' concert' }
     location { Faker::Address.full_address }
-    happens_at { Faker::Time.forward(days: 365, format: :long) }
+    happens_at { Faker::Time.forward(days: 365, format: :short) }
     ticket_price { Faker::Number.within(range: 1..100_000) }
     tickets_available { Faker::Number.within(range: 1..100_000) }
   end
